Fixed all the new compiler warnings.

+++ Project Euler/Files/Classes/Question Objects/Questions 1-100/Questions 1-25/Questions 6-10/Question8.m
@@ -247,7 +247,7 @@
         // Set the MovingProduct5 to the returned value of this method based on
         // the current index + 1 (i.e.: the "digit" to the right of the current
         // one).
-        movingProduct = [self string:aString index:(subStringRange.location + 1)];
+        movingProduct = [self string:aString index:((uint)subStringRange.location + 1)];
         
         // Exit the loop.
         break;
